refac general:  cast and astpe
