+++ ivy/functional/backends/numpy/data_type.py
@@ -56,3 +56,11 @@ def result_type(*arrays_and_dtypes: Union[np.ndarray, np.dtype]) -> np.dtype:
   
 def broadcast_to(x: np.ndarray, shape: Tuple[int,...]) -> np.ndarray:
     return np.broadcast_to(x,shape)
+
+
+
+def cast(x, dtype):
+    return x.astype(ivy.dtype_from_str(dtype))
+
+
+astype = cast
