+++ ivy/functional/backends/numpy/old/general.py
@@ -83,12 +83,6 @@ def dtype_bits(dtype_in):
 
 
 
-def cast(x, dtype):
-    return x.astype(dtype_from_str(dtype))
-
-
-astype = cast
-
